Changed route for Donate and Project Details page
