+++ pages/index.tsx
@@ -1,12 +1,22 @@
-import Footer from '../src/features/common/Footer';
 import Layout from '../src/features/common/Layout';
-import HomePage from '../src/features/public/Home';
+import DonateComponent from '../src/features/public/Donations';
+import { context } from '../src/utils/config';
 
-export default function Home() {
+export default function Donate({ projects }: any) {
+  const DonateProps = {
+    projects: projects,
+  };
   return (
     <Layout>
-      <HomePage />
-      <Footer />
+      <DonateComponent {...DonateProps} />
     </Layout>
   );
+}
+
+export async function getStaticProps() {
+  const res = await fetch(`${context.api_url}/app/projects?_scope=map`);
+  const projects = await res.json();
+  return {
+    props: { projects }, // will be passed to the page component as props
+  };
 }
