+++ src/features/public/Donations/components/ProjectSnippet.tsx
@@ -39,11 +39,7 @@ export default function ProjectSnippet({ project }: Props): ReactElement {
       >
         <TreeDonation project={projectDetails} onClose={handleClose} />
       </Modal>
-      <Link
-        prefetch={false}
-        href="/donate/[id]"
-        as={`/donate/${project.properties.id}`}
-      >
+      <Link prefetch={false} href="/[id]" as={`/${project.properties.id}`}>
         <a>
           <div className={styles.projectImage}>
             {project.properties.image &&
